Added Ruggedness and Plotting (#6)
* added ruggedness and plotting

* Modified persistence functions such that it returns persistence over time (over epochs) to support implementation of persistence area under curve, implemented persistence area under curve

* added statsmodel to requirements.txt

* Fixed bug with persistence calculation, updated corresponding tests

* fixed wrong import

* soled the conflict.

* solved the 2 times defined acc list conflict.
