+++ requirements.txt
@@ -29,3 +29,6 @@ sh              # for running bash commands in some tests
 torch==1.11.0
 torchmetrics>=0.7.0
 torchvision==0.12.0
+
+# -------- statistics -------- #
+statsmodels
